+++ persistence/debug.go
@@ -49,7 +49,7 @@ func (m *PersistenceModule) clearState(_ *debug.DebugMessage) {
 		log.Printf("Error creating new context: %s \n", err)
 		return
 	}
-	if err := context.(PostgresContext).DebugClearAll(); err != nil {
+	if err := context.(*PostgresContext).DebugClearAll(); err != nil {
 		log.Printf("Error clearing state: %s \n", err)
 		return
 	}