+++ shared/codec/codec.go
@@ -5,7 +5,7 @@ import (
 	"google.golang.org/protobuf/types/known/anypb"
 )
 
-type Codec interface { // TODO (Team) move to shared. Possibly rename
+type Codec interface {
 	Marshal(proto.Message) ([]byte, error)
 	Unmarshal([]byte, proto.Message) error
 	ToAny(proto.Message) (*anypb.Any, error)
@@ -48,7 +48,9 @@ func (p *ProtoCodec) FromAny(any *anypb.Any) (proto.Message, error) {
 	return msg, nil
 }
 
-// TODO(andrew): Retrieve this from the utility module via the application specific bus
+// DISCUSS: Retrieve this from the utility module via the application specific bus?
+// There are some parts of the code that does not have access to the bus;
+// Example: txIndexer
 func GetCodec() Codec {
 	return &ProtoCodec{}
 }